requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
email_validator
flask_wtf
pillow
Flask-Mail
itsdangerous